make code cross arch. long->int32, short->int16

+++ src/gusanos/2xsai.cpp
@@ -218,7 +218,7 @@ void Super2xSaI_ex(uint8_t *src, uint32_t src_pitch, uint8_t *unused, ALLEGRO_BI
 	int j, v;
 	unsigned int x, y;
 	int sbpp = BYTES_PER_PIXEL(bitmap_color_depth(dest));
-	unsigned long color[16];
+	Uint32 color[16];
 	unsigned char **src_line = new unsigned char*[4];
 	unsigned char **dst_line = new unsigned char*[2];
 
@@ -248,23 +248,23 @@ void Super2xSaI_ex(uint8_t *src, uint32_t src_pitch, uint8_t *unused, ALLEGRO_BI
 	x = 0, y = 0;
 	
 	if (PixelsPerMask == 2) {
-		unsigned short *sbp;
-		sbp = (unsigned short*)src_line[0];
+		Uint16 *sbp;
+		sbp = (Uint16*)src_line[0];
 		color[0] = *sbp;       color[1] = color[0];   color[2] = color[0];    color[3] = color[0];
 		color[4] = color[0];   color[5] = color[0];   color[6] = *(sbp + 1);  color[7] = *(sbp + 2);
-		sbp = (unsigned short*)src_line[2];
+		sbp = (Uint16*)src_line[2];
 		color[8] = *sbp;     color[9] = color[8];     color[10] = *(sbp + 1); color[11] = *(sbp + 2);
-		sbp = (unsigned short*)src_line[3];
+		sbp = (Uint16*)src_line[3];
 		color[12] = *sbp;    color[13] = color[12];   color[14] = *(sbp + 1); color[15] = *(sbp + 2);
 	}
 	else {
-		unsigned long *lbp;
-		lbp = (unsigned long*)src_line[0];
+		Uint32 *lbp;
+		lbp = (Uint32*)src_line[0];
 		color[0] = *lbp;       color[1] = color[0];   color[2] = color[0];    color[3] = color[0];
 		color[4] = color[0];   color[5] = color[0];   color[6] = *(lbp + 1);  color[7] = *(lbp + 2);
-		lbp = (unsigned long*)src_line[2];
+		lbp = (Uint32*)src_line[2];
 		color[8] = *lbp;     color[9] = color[8];     color[10] = *(lbp + 1); color[11] = *(lbp + 2);
-		lbp = (unsigned long*)src_line[3];
+		lbp = (Uint32*)src_line[3];
 		color[12] = *lbp;    color[13] = color[12];   color[14] = *(lbp + 1); color[15] = *(lbp + 2);
 	}
 
@@ -273,7 +273,7 @@ void Super2xSaI_ex(uint8_t *src, uint32_t src_pitch, uint8_t *unused, ALLEGRO_BI
 		/* Todo: x = width - 2, x = width - 1 */
 		
 		for (x = 0; x < width; x++) {
-			unsigned long product1a, product1b, product2a, product2b;
+			Uint32 product1a, product1b, product2a, product2b;
 
 //---------------------------------------  B0 B1 B2 B3    0  1  2  3
 //                                         4  5* 6  S2 -> 4  5* 6  7
@@ -337,14 +337,14 @@ void Super2xSaI_ex(uint8_t *src, uint32_t src_pitch, uint8_t *unused, ALLEGRO_BI
 				product1a = color[5];
 	
 			if (PixelsPerMask == 2) {
-				*((unsigned long *) (&dst_line[0][x * 4])) = product1a | (product1b << 16);
-				*((unsigned long *) (&dst_line[1][x * 4])) = product2a | (product2b << 16);
+				*((Uint32 *) (&dst_line[0][x * 4])) = product1a | (product1b << 16);
+				*((Uint32 *) (&dst_line[1][x * 4])) = product2a | (product2b << 16);
 			}
 			else {
-				*((unsigned long *) (&dst_line[0][x * 8])) = product1a;
-				*((unsigned long *) (&dst_line[0][x * 8 + 4])) = product1b;
-				*((unsigned long *) (&dst_line[1][x * 8])) = product2a;
-				*((unsigned long *) (&dst_line[1][x * 8 + 4])) = product2b;
+				*((Uint32 *) (&dst_line[0][x * 8])) = product1a;
+				*((Uint32 *) (&dst_line[0][x * 8 + 4])) = product1b;
+				*((Uint32 *) (&dst_line[1][x * 8])) = product2a;
+				*((Uint32 *) (&dst_line[1][x * 8 + 4])) = product2b;
 			}
 			
 			/* Move color matrix forward */
@@ -355,16 +355,16 @@ void Super2xSaI_ex(uint8_t *src, uint32_t src_pitch, uint8_t *unused, ALLEGRO_BI
 			if (x < width - 3) {
 				x += 3;
 				if (PixelsPerMask == 2) {
-					color[3] = *(((unsigned short*)src_line[0]) + x);					
-					color[7] = *(((unsigned short*)src_line[1]) + x);
-					color[11] = *(((unsigned short*)src_line[2]) + x);
-					color[15] = *(((unsigned short*)src_line[3]) + x);
+					color[3] = *(((Uint16*)src_line[0]) + x);					
+					color[7] = *(((Uint16*)src_line[1]) + x);
+					color[11] = *(((Uint16*)src_line[2]) + x);
+					color[15] = *(((Uint16*)src_line[3]) + x);
 				}
 				else {
-					color[3] = *(((unsigned long*)src_line[0]) + x);
-					color[7] = *(((unsigned long*)src_line[1]) + x);
-					color[11] = *(((unsigned long*)src_line[2]) + x);
-					color[15] = *(((unsigned long*)src_line[3]) + x);
+					color[3] = *(((Uint32*)src_line[0]) + x);
+					color[7] = *(((Uint32*)src_line[1]) + x);
+					color[11] = *(((Uint32*)src_line[2]) + x);
+					color[15] = *(((Uint32*)src_line[3]) + x);
 				}
 				x -= 3;
 			}
@@ -383,41 +383,41 @@ void Super2xSaI_ex(uint8_t *src, uint32_t src_pitch, uint8_t *unused, ALLEGRO_BI
 			
 		/* Then shift the color matrix up */
 		if (PixelsPerMask == 2) {
-			unsigned short *sbp;
-			sbp = (unsigned short*)src_line[0];
+			Uint16 *sbp;
+			sbp = (Uint16*)src_line[0];
 			color[0] = *sbp;     color[1] = color[0];    color[2] = *(sbp + 1);  color[3] = *(sbp + 2);
-			sbp = (unsigned short*)src_line[1];
+			sbp = (Uint16*)src_line[1];
 			color[4] = *sbp;     color[5] = color[4];    color[6] = *(sbp + 1);  color[7] = *(sbp + 2);
-			sbp = (unsigned short*)src_line[2];
+			sbp = (Uint16*)src_line[2];
 			color[8] = *sbp;     color[9] = color[9];    color[10] = *(sbp + 1); color[11] = *(sbp + 2);
-			sbp = (unsigned short*)src_line[3];
+			sbp = (Uint16*)src_line[3];
 			color[12] = *sbp;    color[13] = color[12];  color[14] = *(sbp + 1); color[15] = *(sbp + 2);
 		}
 		else {
-			unsigned long *lbp;
-			lbp = (unsigned long*)src_line[0];
+			Uint32 *lbp;
+			lbp = (Uint32*)src_line[0];
 			color[0] = *lbp;     color[1] = color[0];    color[2] = *(lbp + 1);  color[3] = *(lbp + 2);
-			lbp = (unsigned long*)src_line[1];
+			lbp = (Uint32*)src_line[1];
 			color[4] = *lbp;     color[5] = color[4];    color[6] = *(lbp + 1);  color[7] = *(lbp + 2);
-			lbp = (unsigned long*)src_line[2];
+			lbp = (Uint32*)src_line[2];
 			color[8] = *lbp;     color[9] = color[9];    color[10] = *(lbp + 1); color[11] = *(lbp + 2);
-			lbp = (unsigned long*)src_line[3];
+			lbp = (Uint32*)src_line[3];
 			color[12] = *lbp;    color[13] = color[12];  color[14] = *(lbp + 1); color[15] = *(lbp + 2);
 		}
 		
 		
 		/* Write the 2 lines, if not already done so */
 		if (v) {
-			unsigned long dst_addr;
+			Uint32 dst_addr;
 		
 			dst_addr = bmp_write_line(dest, y * 2);
-			for (j = 0; j < dest->w * sbpp; (j += sizeof(long)), (dst_addr += sizeof(long)))
-				bmp_write32(dst_addr, *((unsigned long *) (dst_line[0] + j)));
+			for (j = 0; j < dest->w * sbpp; (j += sizeof(Uint32)), (dst_addr += sizeof(Uint32)))
+				bmp_write32(dst_addr, *((Uint32 *) (dst_line[0] + j)));
 			
 			
 			dst_addr = bmp_write_line(dest, y * 2 + 1);
-			for (j = 0; j < dest->w * sbpp; (j += sizeof(long)), (dst_addr += sizeof(long)))
-				bmp_write32(dst_addr, *((unsigned long *) (dst_line[1] + j)));
+			for (j = 0; j < dest->w * sbpp; (j += sizeof(Uint32)), (dst_addr += sizeof(Uint32)))
+				bmp_write32(dst_addr, *((Uint32 *) (dst_line[1] + j)));
 		}
 		else {
 			if (y < height - 1) {
@@ -479,7 +479,7 @@ void SuperEagle_ex(uint8_t *src, uint32_t src_pitch, uint8_t *unused, ALLEGRO_BI
 	int j, v;
 	unsigned int x, y;
 	int sbpp = BYTES_PER_PIXEL(bitmap_color_depth(dest));
-	unsigned long color[12];
+	Uint32 color[12];
 	unsigned char **src_line = new unsigned char*[4];
 	unsigned char **dst_line = new unsigned char*[2];
 
@@ -509,23 +509,23 @@ void SuperEagle_ex(uint8_t *src, uint32_t src_pitch, uint8_t *unused, ALLEGRO_BI
 	x = 0, y = 0;
 	
 	if (PixelsPerMask == 2) {
-		unsigned short *sbp;
-		sbp = (unsigned short*)src_line[0];
+		Uint16 *sbp;
+		sbp = (Uint16*)src_line[0];
 		color[0] = *sbp;       color[1] = color[0];   color[2] = color[0];    color[3] = color[0];
 		color[4] = *(sbp + 1); color[5] = *(sbp + 2);
-		sbp = (unsigned short*)src_line[2];
+		sbp = (Uint16*)src_line[2];
 		color[6] = *sbp;     color[7] = color[6];     color[8] = *(sbp + 1); color[9] = *(sbp + 2);
-		sbp = (unsigned short*)src_line[3];
+		sbp = (Uint16*)src_line[3];
 		color[10] = *sbp;    color[11] = *(sbp + 1); 
 	}
 	else {
-		unsigned long *lbp;
-		lbp = (unsigned long*)src_line[0];
+		Uint32 *lbp;
+		lbp = (Uint32*)src_line[0];
 		color[0] = *lbp;       color[1] = color[0];   color[2] = color[0];    color[3] = color[0];
 		color[4] = *(lbp + 1); color[5] = *(lbp + 2);
-		lbp = (unsigned long*)src_line[2];
+		lbp = (Uint32*)src_line[2];
 		color[6] = *lbp;     color[7] = color[6];     color[8] = *(lbp + 1); color[9] = *(lbp + 2);
-		lbp = (unsigned long*)src_line[3];
+		lbp = (Uint32*)src_line[3];
 		color[10] = *lbp;    color[11] = *(lbp + 1);
 	}
 
@@ -534,7 +534,7 @@ void SuperEagle_ex(uint8_t *src, uint32_t src_pitch, uint8_t *unused, ALLEGRO_BI
 		/* Todo: x = width - 2, x = width - 1 */
 		
 		for (x = 0; x < width; x++) {
-			unsigned long product1a, product1b, product2a, product2b;
+			Uint32 product1a, product1b, product2a, product2b;
 
 //---------------------------------------     B1 B2           0  1
 //                                         4  5  6  S2 ->  2  3  4  5
@@ -600,14 +600,14 @@ void SuperEagle_ex(uint8_t *src, uint32_t src_pitch, uint8_t *unused, ALLEGRO_BI
 			}
 
 			if (PixelsPerMask == 2) {
-				*((unsigned long *) (&dst_line[0][x * 4])) = product1a | (product1b << 16);
-				*((unsigned long *) (&dst_line[1][x * 4])) = product2a | (product2b << 16);
+				*((Uint32 *) (&dst_line[0][x * 4])) = product1a | (product1b << 16);
+				*((Uint32 *) (&dst_line[1][x * 4])) = product2a | (product2b << 16);
 			}
 			else {
-				*((unsigned long *) (&dst_line[0][x * 8])) = product1a;
-				*((unsigned long *) (&dst_line[0][x * 8 + 4])) = product1b;
-				*((unsigned long *) (&dst_line[1][x * 8])) = product2a;
-				*((unsigned long *) (&dst_line[1][x * 8 + 4])) = product2b;
+				*((Uint32 *) (&dst_line[0][x * 8])) = product1a;
+				*((Uint32 *) (&dst_line[0][x * 8 + 4])) = product1b;
+				*((Uint32 *) (&dst_line[1][x * 8])) = product2a;
+				*((Uint32 *) (&dst_line[1][x * 8 + 4])) = product2b;
 			}
 			
 			/* Move color matrix forward */
@@ -619,20 +619,20 @@ void SuperEagle_ex(uint8_t *src, uint32_t src_pitch, uint8_t *unused, ALLEGRO_BI
 			if (x < width - 2) {
 				x += 2;
 				if (PixelsPerMask == 2) {
-					color[1] = *(((unsigned short*)src_line[0]) + x);
+					color[1] = *(((Uint16*)src_line[0]) + x);
 					if (x < width) {
-						color[5] = *(((unsigned short*)src_line[1]) + x + 1);
-						color[9] = *(((unsigned short*)src_line[2]) + x + 1);
+						color[5] = *(((Uint16*)src_line[1]) + x + 1);
+						color[9] = *(((Uint16*)src_line[2]) + x + 1);
 					}
-					color[11] = *(((unsigned short*)src_line[3]) + x);
+					color[11] = *(((Uint16*)src_line[3]) + x);
 				}
 				else {
-					color[1] = *(((unsigned long*)src_line[0]) + x);
+					color[1] = *(((Uint32*)src_line[0]) + x);
 					if (x < width) {
-						color[5] = *(((unsigned long*)src_line[1]) + x + 1);
-						color[9] = *(((unsigned long*)src_line[2]) + x + 1);
+						color[5] = *(((Uint32*)src_line[1]) + x + 1);
+						color[9] = *(((Uint32*)src_line[2]) + x + 1);
 					}
-					color[11] = *(((unsigned long*)src_line[3]) + x);
+					color[11] = *(((Uint32*)src_line[3]) + x);
 				}
 				x -= 2;
 			}
@@ -651,40 +651,40 @@ void SuperEagle_ex(uint8_t *src, uint32_t src_pitch, uint8_t *unused, ALLEGRO_BI
 			
 		/* Then shift the color matrix up */
 		if (PixelsPerMask == 2) {
-			unsigned short *sbp;
-			sbp = (unsigned short*)src_line[0];
+			Uint16 *sbp;
+			sbp = (Uint16*)src_line[0];
 			color[0] = *sbp;     color[1] = *(sbp + 1);
-			sbp = (unsigned short*)src_line[1];
+			sbp = (Uint16*)src_line[1];
 			color[2] = *sbp;     color[3] = color[2];    color[4] = *(sbp + 1);  color[5] = *(sbp + 2);
-			sbp = (unsigned short*)src_line[2];
+			sbp = (Uint16*)src_line[2];
 			color[6] = *sbp;     color[7] = color[6];    color[8] = *(sbp + 1);  color[9] = *(sbp + 2);
-			sbp = (unsigned short*)src_line[3];
+			sbp = (Uint16*)src_line[3];
 			color[10] = *sbp;    color[11] = *(sbp + 1);
 		}
 		else {
-			unsigned long *lbp;
-			lbp = (unsigned long*)src_line[0];
+			Uint32 *lbp;
+			lbp = (Uint32*)src_line[0];
 			color[0] = *lbp;     color[1] = *(lbp + 1);
-			lbp = (unsigned long*)src_line[1];
+			lbp = (Uint32*)src_line[1];
 			color[2] = *lbp;     color[3] = color[2];    color[4] = *(lbp + 1);  color[5] = *(lbp + 2);
-			lbp = (unsigned long*)src_line[2];
+			lbp = (Uint32*)src_line[2];
 			color[6] = *lbp;     color[7] = color[6];    color[8] = *(lbp + 1);  color[9] = *(lbp + 2);
-			lbp = (unsigned long*)src_line[3];
+			lbp = (Uint32*)src_line[3];
 			color[10] = *lbp;    color[11] = *(lbp + 1);
 		}
 
 
 		/* Write the 2 lines, if not already done so */
 		if (v) {
-			unsigned long dst_addr;
+			Uint32 dst_addr;
 		
 			dst_addr = bmp_write_line(dest, y * 2);
-			for (j = 0; j < dest->w * sbpp; j += sizeof(long))
-				bmp_write32(dst_addr + j, *((unsigned long *) (dst_line[0] + j)));
+			for (j = 0; j < dest->w * sbpp; j += sizeof(Uint32))
+				bmp_write32(dst_addr + j, *((Uint32 *) (dst_line[0] + j)));
 				
 			dst_addr = bmp_write_line(dest, y * 2 + 1);
-			for (j = 0; j < dest->w * sbpp; j += sizeof(long))
-				bmp_write32(dst_addr + j, *((unsigned long *) (dst_line[1] + j)));
+			for (j = 0; j < dest->w * sbpp; j += sizeof(Uint32))
+				bmp_write32(dst_addr + j, *((Uint32 *) (dst_line[1] + j)));
 		}
 		else {
 			if (y < height - 1) {
